added the registration form using handlebars
