+++ src/app.js
@@ -1,4 +1,6 @@
 const express = require('express');
+const exphbs  = require('express-handlebars');
+const path = require('path');
 require('dotenv').config();
 const app = express();
 
@@ -6,10 +8,17 @@ const PORT = process.env.PORT || 8080;
 app.listen(PORT, () =>{
  console.log(`Listining on Port: ${PORT}`)
 });
-app.use(express.json()); 
+
+app.engine('handlebars', exphbs({defaultLayout: 'default'}));
+app.set('views', path.join(__dirname, 'views/'));
+app.set('view engine', 'handlebars');
 
 app.get('/', (req, res)=>{
     res.send('<h1>Hello world<h1>');
 })
 
-app.use('/Dashboard', require('./views/dashboard'));
+app.use(express.static(path.join(__dirname, '../public')));
+
+
+//All routes
+app.use('/Dashboard', require('./routes/Dashboard'));
